feat(adapters): migrate state-machine overrides to behavior packs (6 adapters)
Fan out the declarative behavior DSL across the billing adapters. Each
adapter's lifecycle state machines are now authored as YAML and executed by
the shared interpreter; the hand-written override handlers they replace are
removed. Non-declarative handlers (factory creates, computed/aggregate reads,
list-shaping, marshallers) intentionally remain as code — the escape hatch.

- chargebee:   subscriptions + invoices (7 actions)   25/25 tests
- gocardless:  subscriptions/payments/mandates (7)     14/14 tests
- revenuecat:  subscriptions/products/offerings/ent (8) 14/14 tests
- zuora:       subscriptions + credit-memos (5)         55/55 tests
- lemonsqueezy: subscriptions (2)                       19/19 tests
- paddle:      subscriptions (2)                        14/14 tests

Full workspace: 24/24 builds, 45/45 test tasks green.
