+++ packages/adapters/adapter-chargebee/package.json
@@ -20,7 +20,8 @@
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
-    "generate": "tsx scripts/chargebee-codegen.ts",
+    "generate": "tsx scripts/chargebee-codegen.ts && mimic-behavior-codegen",
+    "generate:behavior": "mimic-behavior-codegen",
     "test": "vitest run",
     "typecheck": "tsc --noEmit",
     "clean": "rm -rf dist"
