+++ packages/adapters/adapter-chargebee/src/overrides/subscriptions.ts
@@ -1,7 +1,6 @@
 import type { StateStore } from '@mimicai/core';
 import type { OverrideHandler } from '@mimicai/adapter-sdk';
 import { unixNow, generateId } from '@mimicai/adapter-sdk';
-import { chargebeeNotFound, chargebeeStateError } from '../chargebee-errors.js';
 import { SCHEMA_DEFAULTS } from '../generated/schemas.js';
 
 const NS = 'chargebee:subscriptions';
@@ -28,126 +27,10 @@ export function buildCreateSubscriptionHandler(store: StateStore): OverrideHandl
   };
 }
 
-export function buildCancelHandler(store: StateStore): OverrideHandler {
-  return async (req, reply) => {
-    const id = (req.params as Record<string, string>)['subscription_id'];
-    const existing = store.get<Record<string, unknown>>(NS, id!);
-
-    if (!existing) {
-      return reply.code(404).send(chargebeeNotFound('subscriptions', id!));
-    }
-
-    const status = existing.status as string;
-    if (status === 'cancelled' || status === 'non_renewing') {
-      return reply.code(400).send(chargebeeStateError(
-        `Subscription ${id} is already ${status}, cannot cancel`,
-      ));
-    }
-
-    const body = (req.body ?? {}) as Record<string, unknown>;
-    const endOfTerm = body.end_of_term === true || body.end_of_term === 'true';
-    const now = unixNow();
-
-    const updated: Record<string, unknown> = {
-      ...existing,
-      status: endOfTerm ? 'non_renewing' : 'cancelled',
-      cancel_reason: body.cancel_reason_code ?? 'not_paid',
-      cancelled_at: endOfTerm ? undefined : now,
-      updated_at: now,
-      resource_version: now * 1000,
-    };
-
-    store.set(NS, id!, updated);
-    return reply.code(200).send(updated);
-  };
-}
-
-export function buildReactivateHandler(store: StateStore): OverrideHandler {
-  return async (req, reply) => {
-    const id = (req.params as Record<string, string>)['subscription_id'];
-    const existing = store.get<Record<string, unknown>>(NS, id!);
-
-    if (!existing) {
-      return reply.code(404).send(chargebeeNotFound('subscriptions', id!));
-    }
-
-    if (existing.status !== 'cancelled' && existing.status !== 'non_renewing') {
-      return reply.code(400).send(chargebeeStateError(
-        `Subscription ${id} is ${existing.status}, cannot reactivate`,
-      ));
-    }
-
-    const now = unixNow();
-    const updated: Record<string, unknown> = {
-      ...existing,
-      status: 'active',
-      cancelled_at: null,
-      cancel_reason: null,
-      activated_at: now,
-      updated_at: now,
-      resource_version: now * 1000,
-    };
-
-    store.set(NS, id!, updated);
-    return reply.code(200).send(updated);
-  };
-}
-
-export function buildPauseHandler(store: StateStore): OverrideHandler {
-  return async (req, reply) => {
-    const id = (req.params as Record<string, string>)['subscription_id'];
-    const existing = store.get<Record<string, unknown>>(NS, id!);
-
-    if (!existing) {
-      return reply.code(404).send(chargebeeNotFound('subscriptions', id!));
-    }
-
-    if (existing.status !== 'active') {
-      return reply.code(400).send(chargebeeStateError(
-        `Subscription ${id} is ${existing.status}, cannot pause`,
-      ));
-    }
-
-    const now = unixNow();
-    const updated: Record<string, unknown> = {
-      ...existing,
-      status: 'paused',
-      pause_date: now,
-      updated_at: now,
-      resource_version: now * 1000,
-    };
-
-    store.set(NS, id!, updated);
-    return reply.code(200).send(updated);
-  };
-}
-
-export function buildResumeHandler(store: StateStore): OverrideHandler {
-  return async (req, reply) => {
-    const id = (req.params as Record<string, string>)['subscription_id'];
-    const existing = store.get<Record<string, unknown>>(NS, id!);
-
-    if (!existing) {
-      return reply.code(404).send(chargebeeNotFound('subscriptions', id!));
-    }
-
-    if (existing.status !== 'paused') {
-      return reply.code(400).send(chargebeeStateError(
-        `Subscription ${id} is ${existing.status}, cannot resume`,
-      ));
-    }
-
-    const now = unixNow();
-    const updated: Record<string, unknown> = {
-      ...existing,
-      status: 'active',
-      pause_date: null,
-      resume_date: now,
-      updated_at: now,
-      resource_version: now * 1000,
-    };
-
-    store.set(NS, id!, updated);
-    return reply.code(200).send(updated);
-  };
-}
+// ---------------------------------------------------------------------------
+// The lifecycle state machines (cancel / reactivate / pause / resume) were
+// migrated to the declarative behavior pack at src/behavior/subscriptions.yaml
+// and are now mounted via mountBehaviorPacks. Only the creation handler — which
+// relies on the schema-default factory and response pre-wrapping — remains here
+// as a documented escape hatch.
+// ---------------------------------------------------------------------------
