+++ packages/adapters/adapter-paddle/src/paddle-adapter.ts
@@ -16,6 +16,8 @@ import type { GeneratedRoute } from './generated/routes.js';
 
 import * as subscriptionOverrides from './overrides/subscriptions.js';
 import * as transactionOverrides from './overrides/transactions.js';
+import { behaviorPacks } from './generated/behavior.js';
+import { paddleError, paddleStateError } from './paddle-errors.js';
 
 function ns(resource: string): string {
   return `paddle:${resource}`;
@@ -159,18 +161,23 @@ export class PaddleAdapter extends OpenApiMockAdapter<PaddleConfig> {
   // ---------------------------------------------------------------------------
 
   private mountOverrides(store: StateStore): void {
-    // Subscription lifecycle
+    // Subscription lifecycle (declarative behavior pack): activate, resume.
+    // cancel & pause stay hand-written — their next_billing_period branch
+    // computes an ISO timestamp 30 days out, which the DSL cannot express.
     this.registerOverride('POST', '/subscriptions/:subscription_id/cancel',
       subscriptionOverrides.buildCancelHandler(store));
     this.registerOverride('POST', '/subscriptions/:subscription_id/pause',
       subscriptionOverrides.buildPauseHandler(store));
-    this.registerOverride('POST', '/subscriptions/:subscription_id/resume',
-      subscriptionOverrides.buildResumeHandler(store));
-    this.registerOverride('POST', '/subscriptions/:subscription_id/activate',
-      subscriptionOverrides.buildActivateHandler(store));
 
-    // Transaction lifecycle
+    // Transaction lifecycle (whole-body merge marshaller — kept as override).
     this.registerOverride('POST', '/transactions/:transaction_id/revise',
       transactionOverrides.buildReviseHandler(store));
+
+    // Declarative behavior packs (activate, resume).
+    this.mountBehaviorPacks(store, behaviorPacks, (e) =>
+      e.kind === 'not_found'
+        ? paddleError('not_found', e.message)
+        : paddleStateError(e.message),
+    );
   }
 }
