+++ packages/adapters/adapter-zuora/package.json
@@ -21,7 +21,8 @@
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
-    "generate": "tsx scripts/zuora-codegen.ts",
+    "generate": "tsx scripts/zuora-codegen.ts && mimic-behavior-codegen",
+    "generate:behavior": "mimic-behavior-codegen",
     "test": "vitest run",
     "typecheck": "tsc --noEmit",
     "clean": "rm -rf dist"
